requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
pyTelegramBotAPI==4.7.1
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
